Divide up C elements interface. We want to separate the environment functions, needed for jet bindings, from the execution functions.
